// desktop/src/open-api/model/customFieldType.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type CustomFieldType = 'TEXT' | 'DATE' | 'SELECT' | 'CURRENCY' | 'BOOLEAN';

export const CustomFieldType = {
    Text: 'TEXT' as CustomFieldType,
    Date: 'DATE' as CustomFieldType,
    Select: 'SELECT' as CustomFieldType,
    Currency: 'CURRENCY' as CustomFieldType,
    Boolean: 'BOOLEAN' as CustomFieldType
};